M-13: Рефакторинг кода
