+++ app/realty/models/floor_model.py
@@ -1,0 +1,6 @@
+from django.db import models
+
+
+class Floor(models.Model):
+    name = models.CharField(max_length=100)
+    number = models.PositiveSmallIntegerField(db_index=True)
